Ensure foreignKey Properties and Made remove Caregiver

+++ HealthDevice/DTO/DashBoard.cs
@@ -6,6 +6,6 @@ public class DashBoard
     public int HeartRate { get; set; }
     public float SpO2 { get; set; }
     public int steps { get; set; }
-    public double distance { get; set; }
+    public float distance { get; set; }
     public int allFall { get; set; }
 }
