+++ HealthDevice/Services/GeoService.cs
@@ -87,7 +87,7 @@ public class GeoService : IGeoService
         return formattedAddress;
     }
 
-    public static double CalculateDistance(Location locationA, Location locationB)
+    public static float CalculateDistance(Location locationA, Location locationB)
     {
         double dLat = (locationA.Latitude - locationB.Latitude) * Math.PI / 180;
         double dLon = (locationA.Longitude - locationB.Longitude) * Math.PI / 180;
@@ -98,7 +98,7 @@ public class GeoService : IGeoService
                    Math.Cos(lat1) * Math.Cos(lat2) *
                    Math.Pow(Math.Sin(dLon / 2), 2);
         double c = 2 * Math.Atan2(Math.Sqrt(a), Math.Sqrt(1 - a));
-        double d = 6371 * c;
+        float d = (float)(6371 * c);
         
         return d; // Distance in kilometers
     }
